requirements.txt
gunicorn
Flask
line-bot-sdk
requests
feedparser
pymongo
beautifulsoup4